requirements.txt
asgiref==3.8.1
certifi==2024.2.2
chardet==4.0.0
Django==3.2
djangorestframework==3.12.4
idna==2.10
pytz==2024.1
requests==2.25.1
setuptools==69.2.0
sqlparse==0.4.4
urllib3==1.26.18
